services/api/requirements.txt
fastapi==0.105.0
gunicorn==21.2.0
uvicorn==0.25.0
langchain==0.1.0
langchainhub==0.1.14
openai==1.6.1
langchain-openai==0.0.2
python-multipart==0.0.6
aiohttp==3.9.1
pinecone-client==2.2.4
